Revert "react todo list connected to mongo database."
