+++ client/src/App.js
@@ -9,7 +9,7 @@ import Dashboard from "./pages/Dashboard";
 import Registration from "./pages/Registration.js";
 import SignIn from "./pages/SignIn.js";
 import FaceRecognition from "./pages/FaceRecognition.js";
-import TodoPage from "./pages/TodoPage/";
+import TodoPage from "./pages/TodoPage.js";
 
 
 // import { Helmet } from 'react-helmet';
