--- client/src/components/DeleteBtn/index.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-
-// The ...props means, spread all of the passed props onto this element
-// That way we don't have to define them all individually
-function DeleteBtn(props) {
-  return (
-    <span className="delete-btn" {...props} role="button" tabIndex="0">
-      ✗
-    </span>
-  );
-}
-
-export default DeleteBtn;
