--- client/src/components/Form/TextArea.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export function TextArea(props) {
-  return (
-    <div className="form-group">
-      <textarea className="form-control" rows="20" {...props} />
-    </div>
-  );
-}
