+++ client/src/pages/TodoPage.js
@@ -1,0 +1,23 @@
+import React, { Component } from "react";
+import { Link } from "react-router-dom";
+import { Col, Row, Container } from "../components/Grid";
+
+class TodoPage extends Component {
+
+    render() {
+        return (
+            <div>
+                <Container>
+                    <Row>
+                        <Col size="md-12">
+
+                        </Col>
+                    </Row>
+                </Container>
+            </div>
+        );
+    }
+}
+
+
+export default TodoPage;
