--- client/src/pages/TodoPage/index.js
@@ -1,1 +1,0 @@
-export { default } from "./TodoPage.js";
