--- client/src/utils/API.js
@@ -1,20 +1,0 @@
-import axios from "axios";
-
-export default {
-  // Gets all todos
-  getToDos: function() {
-    return axios.get("/api/todos");
-  },
-  // Gets the todo with the given id
-  getToDo: function(id) {
-    return axios.get("/api/todos/" + id);
-  },
-  // Deletes the todo with the given id
-  deleteToDo: function(id) {
-    return axios.delete("/api/todos/" + id);
-  },
-  // Saves a todo to the database
-  saveToDo: function(todoData) {
-    return axios.post("/api/todos", todoData);
-  }
-};
