+++ controllers/todosController.js
@@ -1,6 +1,6 @@
 const db = require("../models");
 
-// Defining methods for the todosController
+// Defining methods for the booksController
 module.exports = {
   findAll: function (req, res) {
     db.Todo
