+++ models/todo.js
@@ -13,10 +13,6 @@ const todoSchema = new Schema({
   date: {
     type: Date,
     default: Date.now
-  },
-  completed: {
-    type: Boolean,
-    default: false
   }
 });
