+++ routes/api/todos.js
@@ -1,12 +1,12 @@
 const router = require("express").Router();
 const todosController = require("../../controllers/todosController");
 
-// Matches with "/api/todos"
+// Matches with "/api/books"
 router.route("/")
   .get(todosController.findAll)
   .post(todosController.create);
 
-// Matches with "/api/todos/:id"
+// Matches with "/api/books/:id"
 router
   .route("/:id")
   .get(todosController.findById)
