+++ server.js
@@ -17,16 +17,13 @@ const PORT = process.env.PORT || 3001;
   // ? CORs config
   app.use(cors());
 
-// * Custom API Router config for todo list
-const routes = require("./routes");
+// * Router config
+// const routes = require("./routes")(app);
 
 // * Server config cont.
 if (process.env.NODE_ENV === "production") {
   app.use(express.static("client/build"));
 }
-
-// Add routes, both API and view
-app.use(routes);
 
 // * Serving up our static content
 app.use(express.static("public"))
@@ -49,12 +46,6 @@ module.exports = app;
 // const axios = require("axios");
 // var passport = require("./config/passport");
 // var faceapi = require("face-api.js");
-
-// $ heroku config
-// === calm-woodland-49321 Config Vars
-// MONGODB_URI: mongodb://heroku_tmwk490b:urd1rji0j9392e0748uf710hc7@ds333238.mlab.com:33238/heroku_tmwk490b
-
-
 
 //mlab
 // mongodb:
